Change the participant service to depend only on tx_manager and hackathon_service
