+++ services/py-api/src/server/handlers/participants_handlers.py
@@ -5,12 +5,11 @@ from starlette.responses import Response
 from src.server.exception import DuplicateEmailError, DuplicateTeamNameError, HackathonCapacityExceededError
 from src.server.schemas.request_schemas.schemas import ParticipantRequestBody
 from src.server.schemas.response_schemas.schemas import ParticipantRegisteredInTeamResponse, ErrResponse
-from src.service.participants_service import ParticipantService
-from src.server.exception import DuplicateEmailError, DuplicateTeamNameError, HackathonCapacityExceededError
+from src.service.participants_registration_service import ParticipantRegistrationService
 
 
 class ParticipantHandlers:
-    def __init__(self, service: ParticipantService) -> None:
+    def __init__(self, service: ParticipantRegistrationService) -> None:
         self._service = service
 
     async def create_participant(
