+++ services/py-api/src/server/routes/participants_routes.py
@@ -8,7 +8,8 @@ from src.database.transaction_manager import TransactionManager
 from src.server.handlers.participants_handlers import ParticipantHandlers
 from src.server.schemas.request_schemas.schemas import ParticipantRequestBody
 from src.server.schemas.response_schemas.schemas import ParticipantRegisteredInTeamResponse, ErrResponse
-from src.service.participants_service import ParticipantService
+from src.service.hackathon_service import HackathonService
+from src.service.participants_registration_service import ParticipantRegistrationService
 
 # https://fastapi.tiangolo.com/tutorial/bigger-applications/#apirouter
 participants_router = APIRouter(prefix="/hackathon/participants")
@@ -30,16 +31,23 @@ def _tx_manager(db_manager: DB_MANAGER) -> TransactionManager:
     return TransactionManager(db_manager)
 
 
-def _service(
+def _h_service(
     p_repo: ParticipantsRepository = Depends(_p_repo),
     t_repo: TeamsRepository = Depends(_t_repo),
     tx_manager: TransactionManager = Depends(_tx_manager),
-) -> ParticipantService:
-    return ParticipantService(p_repo, t_repo, tx_manager)
+) -> HackathonService:
+    return HackathonService(p_repo, t_repo, tx_manager)
 
 
-def _handler(service: ParticipantService = Depends(_service)) -> ParticipantHandlers:
-    return ParticipantHandlers(service)
+def _p_service(
+    h_service: HackathonService = Depends(_h_service),
+    tx_manager: TransactionManager = Depends(_tx_manager),
+) -> ParticipantRegistrationService:
+    return ParticipantRegistrationService(h_service, tx_manager)
+
+
+def _handler(p_service: ParticipantRegistrationService = Depends(_p_service)) -> ParticipantHandlers:
+    return ParticipantHandlers(p_service)
 
 
 # https://fastapi.tiangolo.com/advanced/additional-responses/
